refactor(results): give EPOCH_RE and the ready marker a single owner
The run-key regex was defined twice -- aiperf.operator.results_layout.EPOCH_RE
and a hand-copy in config/resolution/resolvers.py -- with a comment instructing
future readers to keep them identical by hand. They did not stay identical: the
config copy was still ^\d{9,11}$ after the operator moved to ^\d{9,20}$, so
every current operator run directory was misparsed as a local layout. Earlier
in this branch I resynced the value; this removes the ability to drift again.

EPOCH_RE now lives in aiperf.common.results_markers, which is dependency-free
and importable from both config and operator, and every consumer imports it
from there. results_layout also stops restating READY_MARKER_NAME.

Also drops the aiperf.kubernetes.sweep_models re-export shim: FailurePolicy and
ConvergenceConfig were re-exported (the latter via module __getattr__) purely so
K8s callers need not change imports. The circular-import the docstring cited no
longer exists -- nothing under aiperf/config imports FailurePolicy from that
module -- so consumers now import from the owning modules.
