+++ src/aiperf/config/resolution/resolvers.py
@@ -10,12 +10,12 @@ The chain is sync (no event loop at call site) and order-explicit.
 from __future__ import annotations
 
 import os
-import re
 import time
 from pathlib import Path
 from typing import TYPE_CHECKING, Protocol, runtime_checkable
 
 from aiperf.common.aiperf_logger import AIPerfLogger
+from aiperf.common.results_markers import EPOCH_RE
 from aiperf.config.artifacts import OutputDefaults
 from aiperf.config.dataset.resolver import DatasetResolver
 
@@ -23,14 +23,6 @@ if TYPE_CHECKING:
     from aiperf.config.resolution.plan import BenchmarkRun
     from aiperf.config.user_files import RunMeta
 
-# Must stay identical to ``aiperf.operator.results_layout.EPOCH_RE``, which owns
-# the run-key format; it is duplicated rather than imported so the config package
-# stays free of operator/kubernetes imports. 9-20 digits covers legacy
-# epoch-seconds directories (10 digits), fractional-second run keys, and the
-# 16-digit whole-second keys the operator emits with a uid-derived suffix. The
-# earlier 9-11 bound predated the suffix and silently misread every current
-# operator run dir as a local layout.
-_EPOCH_RE = re.compile(r"^\d{9,20}$")
 
 __all__ = [
     "ArtifactDirResolver",
@@ -234,11 +226,11 @@ def _derive_run_meta(artifact_dir: Path) -> RunMeta:
 
     Operator-managed runs use the ``<base>/<ns>/<name>/<epoch>`` layout (see
     ``aiperf.operator.results_layout.run_dir``). When the leaf matches
-    ``_EPOCH_RE`` we treat the parent as the AIPerfJob name and the leaf as
+    ``EPOCH_RE`` we treat the parent as the AIPerfJob name and the leaf as
     the run epoch. Otherwise (local-CLI runs, custom paths) the leaf IS the
     run identifier and we substitute wall-clock seconds for the epoch.
 
-    Using ``_EPOCH_RE`` (not ``str.isdigit``) shrinks the false-positive
+    Using ``EPOCH_RE`` (not ``str.isdigit``) shrinks the false-positive
     surface — e.g. ``/tmp/bench/42`` is correctly treated as a local layout
     rather than a one-day-old operator run.
 
@@ -255,7 +247,7 @@ def _derive_run_meta(artifact_dir: Path) -> RunMeta:
     # / ``results_operator.py`` for runs that predate the operator's
     # epoch-stamped layout. Treat it the same as a numeric epoch so the run
     # metadata reflects the historical sentinel, not wall-clock time.
-    if _EPOCH_RE.match(leaf) or leaf == "legacy":
+    if EPOCH_RE.match(leaf) or leaf == "legacy":
         return RunMeta(
             epoch=leaf,
             job_name=artifact_dir.parent.name,
