+++ src/aiperf/operator/routers/_sweeps_artifacts.py
@@ -11,7 +11,8 @@ from typing import Literal
 from fastapi import APIRouter, HTTPException, Request
 from fastapi.responses import Response, StreamingResponse
 
-from aiperf.operator.results_layout import EPOCH_RE, resolve_sweep_dir
+from aiperf.common.results_markers import EPOCH_RE
+from aiperf.operator.results_layout import resolve_sweep_dir
 from aiperf.operator.routers._path_params import validate_results_path_params
 from aiperf.operator.routers.results_files_io import (
     _list_artifact_files,
